requirements.txt
# Dependencias para las herramientas del agente de consulta Darwin

# Configuración y logging
PyYAML>=6.0
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# OpenSearch
opensearch-py>=2.4.0

# AWS Authentication para OpenSearch
requests-aws4auth>=1.1.2

# Utilidades
typing-extensions>=4.8.0

# Opcional: para desarrollo y testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
